pyproject.toml
[project]
name = "workout-api"
version = "0.1.0"
description = ""
authors = [
    {name = "iandra",email = "iandraav@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-pagination (>=0.15.4,<0.16.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
